OM-46608, RB-32748: Average stats responses do not contain related entity types
Background
The StatApiDTO returned from the /stats api includes a field "relatedEntityType". This conveys the entity type of the commodity response as fetched by the filters.

Problem
This field was not populated for the "/stats/Market" implementation when it should have been, i.e. when the "statistics" field includes a filter with the "relatedEntityType" set.

Solution
Add "related_entity_type" to Stats.proto, StatRecord, which is returned from History component on a GetAveragedEntityStats() grpc call - which is what the "/stats/Market" implementation issues. This field is populated from the "entity_type" column of the database records retrived from the MARKET_STATS_xxx table.