OM-48792: XL API: Reliably mapping UUIDs for action requests.
When retrieving actions for groups/entities, use the UuidMapper and the ApiId utility object, which handles the magic mappings (as well as caching other information). I opened OM 48809 as a follow-up to centralize this kind of scope handling and scope expansion in the API component.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/33023/