OM-45087: Addressed issue from Shai about preferring to use class loader.
Shai pointed out that it is preferable to use the class loader to locate
resources in tests instead of hard-coding paths used by Maven. That of
course is only possible if the files already exist.

There are already a number of tests that refer to resources using the
class loader, so this will enhance uniformity as well.

Modified test code to use the class loader to locate resources, when
possible without modifying the code-under-test, but excluded the
`com.vmturbo.platform` module as this change goes only in `develop`
branch and the latter module will be excluded from build soon.

Also:
- Let IntelliJ IDEA rearrange imports according to the Turbonomic
  coding standard.
- Replaced some anonymous classes with anonymous methods and method
  references to improve readability and performance.