OM-48835: XL AO: Only refreshing READY actions.
Don't refresh the action mode or description for non-ready actions. These should be fixed once we accept the action. This fix is quicker and less expensive than the alternative - continuing to retrieve related entity details for In Progress actions.

Also, not reading the action plan topic from the beginning on restart, because we end up re-reading PLAN action plans which is a problem.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/33129/