OM-48852 : Set mariadb systemd service start timeout to infinity. RB:33065
When we exteranlize the db, if there are many transactions to replay
from the redo log, it would take some time for mariadb to startup. If
the systemd startup time is low, it will try to stop/kill mariadb. To
prevent this from happening, set the start and stop timeout to infinity.

See mariadb docs: https://mariadb.com/kb/en/library/systemd/#starting-the-mariadb-server-process