OM-48857: XL API: For Provision actions, set ActionApiDTO.newEntity to provisioned seller.
It makes sense for the "newEntity" in a PROVISION action to be a description of the newly provisioned entity, so we do that. The "target" and "currentEntity" are set to the model seller (the entity to clone from). We also need to make sure we return a description of the newEntity from the repository when construction the action mapping context.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/33154