OM-49143: XL API: Don't set "newEntity" of realtime provision actions.
Distinguish between plans and realtime when mapping actions to ActionApiDTOs and when constructing the ActionSpecMappingContext. Ignore provisioned sellers in realtime. Note - we are now leaving "newEntity" empty. This is fine, and is the same thing that classic does. It's better than setting it to "target"/"currentEntity".

Reviewed At: https://rbcommons.com/s/VMTurbo/r/33211/