OM-48624, RB-33155: 7.21.1: Software Updates screen shows both prior and new component versions after a successful upgrade.
Background
The prior version(s) of XL recorded the build version number for each component in the key/value store on startup. These version numbers, one for each known component type, were returned by the AdminService in response to the "GET /admin/versions" external REST API

Problem
The root path for the component specificy key/value tree was "{component-name}-1", which was a mistake. The instance-id, which was always "-1", was used instead of the component_type. In ClusterMgr this value was extracted from the K/V store and returned as part of the Instance Properties definition. In K8s, however, as horizontal scaling is automatic, the version of all components is the same, and so the version number should be part of the component defaults instead.

Solution
1. in BaseVmtComponent we store the component version number in the key/value store for that component-type instead of the component instance properties.
2. in ClusterMgr we extract the component version number and return it in the component defaults properties
3. in AdminService we fetch the component defaults from ClusterMgr and collection the version numbers from there

Note
Component Configuration, including the component version numbers as handled here, will be undergoing a complete revision. As such, the use of the old-school convention for the root of the key/value store for each component ("{component_type}-1") will need to be revised. In other words, the fix in this review is fairly temporary.