OM-48348 : Connection to VMs filter for storage returns excessive entities. RB:33222
When searching for storage devices which are connected to VMs whose display
name matches a regex, we were returning all storages as the api was
not sending the numberOfHops in the repository search query and
repository ended up returning all storages. This fix
sets the number of hops in the search traversal filter to 1 so that only
direct consumers are returned.