OM-46829 : Remove stale actions re-recommended by market if it was are alredy executed in the previous cycle. RB:33441
The issue is actions which have already been executed show up again.
This is because, if the discovery has not happened and TP broadcasts the
old topology, market will re-generate the same actions.

The fix is to  query the action_history DB table for the last executed
actions in the past 15 minutes and discard the actions which have
already been executed.