OM-49541, RB-33550: ClusterManager: Service restart resets default properties background Componnent Configuration was originally designed to provide for component-type defaults plus instance-specific "overrides". When a component starts up it (1) publishes its default configuration properties to clustermgr, and (2) fetches the "effective" configuration properties from clustermgr, including the defaults plus the overrides.
problem
In moving to kubernetes, the individual component instance id is allocated dynamically by kubernetes, and in the case of horizontal scaling there will be multiple instances of the same component, each with a (different) dynamically generated name. In other words, there is no way to pre-configure one or another component instance.

solution
We will add a "current properties" per-component configuration property as an "override". We also add a DELETE endpoint to that the "current" configuration property may be removed, allowing the "default" to "show through".

In addition - the "node" configuration option, by which a component instance would have affinity to a particular worker node, is completely deprecated and so the UI and fields have been remove.

Further, the original source code was quite old and in need of some cleanup that we did here.