OM-49848, RB-33753: SAML configuration is not passed to the API component
background
The SAML support requires updating, at runtime, a number of configuration items for the API component. In the past these properties were stored in component configuration as "instance properties"

problem
Due to the move to k8s, component instances receive a new instance id whenever the component restarts, and instance properties don't apply when components may be automatically scaled horizontally. As a result, instance properties are being removed from the platform, and so the previous implementation for SAML is not workable.

solution
A new capability was added to configuration, namely component-type-specific "latest" configuration proeprty overrides. This is the correct place for the SAML configuration options to be stored.
To implement this, calls to SAMLKeyValueStore from within api component were replaced with calles to ClusterMgr to set the desired SAML configuration properties.