OM-45627, RB-34305: SaaS Integration with Auth0
Background
We are going to integrate SaaS Authentication with the web service Auth0 as an Identity Provider using the existing SAML integration

Problem
The SAML integration code does not correctly extract the logout token from the SAML Metadata (downloaded from each IDP).
This causes the "SAML logout" handler in API Component not to call the IDP Logout process, thinking that the logout key is not specified.

in particular - for Auth0 the metadata looks like:
<SingleLogoutService Binding="urn:oasis:names...,
where for Okta the tag includes the "md":
<md:SingleLogoutService Binding="urn:oasis:names:tc...

The Single Logout works with Okta, but not Auth0 which was not officially supported during design.

Second Problem
There was a clause inadvertently removed from nginx.conf (OM 47473, I believe) enabling the "/vmturbo/saml" endpoint(s) which are critical to SAML integration

Solution
Provide the "namespace" key to use when extracting the logout token.

Second Solution
Explicitly add back "/vmturbo/saml" as a proxying path.

Note that there were a number of CheckStyle cleanups as well

Testing Done
(1) Set up SAML integration with a test Auth0
(2) go to the URL for the appliance - should be directed to the Auth0 Login;
(3) login to Auth0 - should be directed into the Turbo UI initial panel
(4) click the "logout" icon (lower-left) - should be directed to the Auth0 Login
(5) log in again (perhaps with a different ID, e.g. one with an authorizer role, not administrator) - should be directed to the Turbo UI initial panel (as the other user)