OM-50692 RB:34421 Allow AWS Volume Move action execution
This change is meant to allow AWS Volume Move actions to execute. In order for this to occur, the Volume entity must be treated as the target entity of the action. Assuming the probe capability supports the action, the target entity's settings determine the action's mode. The Move automation setting must be settable for Volume entities, so that a Volume Move action's mode can permit execution.