OM-50985, RB-34318: Component Diagnostic Writer does not stop when the Write Stream closes background Diagnostic Dump files are written using a dynamic ZipOutputStream, and for Diagnostics Download the ZipOutputStream writes directly to the HTTP request write stream.
problem
If the HTTP request write stream closes the error writing to the stream is logged and ignored. For the DiagnosticsWriter writeZipEntry() method that accepts a stream, the entire stream is (attempted to be) written, causing a disasterous stream of error logging.

fix
If there is a error within the stream processing throw an unchecked exception to break out of the stream lambda. Catch that exception and log the message, but don't keep looping.
Also use the container IP address as the consul service address during registration - this will avoid the problem with stale consul service instances causing duplicate diagnostics dumps for the same component instance.