OM-51158, RB-34625: Cluster Manager Container PV ran out of space
background
Uploading diagnostics requires creating a file on a local file system, due to restrictions in the upload-processing application running in the cloud (php?).

problem
This means that, if there are old copies of diagnostics around, or the diagnostics file gets larger than the file space allocated to the ClusterMgr component, then the diagnostics upload will run out of space and fail. Further, if the prior diagnostic upload process is aborted, then there may be an old, partial file left over that would interfere with the next diagnostic upload.

solution
Two parts:
1) create the diagnostic upload file in the /tmp filesystem, which uses Kubernetes Temp Space and so does not have a fixed per-container limit
2) remove all prior diagnostic upload files at the beginning of the process; this will ensure space is available; it will also leave the diagnostics file around in case the upload file is lost - the /tmp file can be retrieved by hand.