OM-39153 RB:34775 fixing simultaneous provisions and suspensions
Problem-1:
A less utilized host has its suspension replayed. The VM on the host undergoes a resize UP. This causes the host's usage to increase and triggers a subsequent activation. Since both these actions occur across 2 sub-cycles, we arent able to collapse the actions.

Problem-2:
A less utilized host in a 3 host cluster has its suspension replayed. However there is a do-not-place-together policy in place on 3 VMs not on this host. After this, we are able to re-activate this suspended host for the VMs only in the 2nd sub-cycle.

Solutions:
For the 1st problem, we replay after resizes.
For the 2nd problem, we move all the possible customers of the suspensionCandidate while replaying and not just the currentCustomers.