OM-51676: Removed redundant calls to getMarket to improve performance.
In the early stages of M2 development, getting the basket bought from a
shopping list required going through the market via a call to
`Economy#getMarket(ShoppingList)`. That was done to save memory by
omitting information available through other means, but it turned out
that the overhead of `Economy#getMarket(ShoppingList)` was too much for
some contexts. Thus a basket field and related getter was added to the
`ShoppingList` class and callers were refactored to use this instead.

Unfortunately, it seems that a few such calls to
`Economy#getMarket(ShoppingList)` survived the initial refactoring and
were found to be responsible for as much as 10% of the time needed for
analysing the DXC topology with SNM enabled.

The call in `Move.updateQuantities` is the important one, but the others
were removed as well for good measure.

Also let IntelliJ IDEA cleanup imports to comply with Turbonomic coding
standard.