OM-49188, RB-34672: XL: Unregister services that may are not cleanly shutdown or after upgrade
background
Components register with Consul so that Clustermgr may track the health of the components, and to request diagnostics dumps. On entry, a component registers with Consul giving the entity type, unique entity id (pod name), and IP. On exit, a component will call "de-register", to keep the list of Components in consul up to date.

problem
If a component does not exit cleanly, it will not get a change to de-register.  Consul keeps old registrations around several days, by default, in case a service comes and goes.
Old services in the list look like they are "down" when you list services, and since they are still on the list the "old" services are contacted for diagnostics dumping - resulting in a (10 second) timeout for each.

solution
There is a feature of the Consul health check which says "after being in critical status for a given time, remove the service from the registry". We are adding a value for this feature, set initially to 1 hour, so any "old" components (pods) will be automatically removed from the Consul list after being "critical" for that length of time.
We also refactored the code to interact with Consul registration into a separate class (ConsulHealthcheckRegistration).