OM-51771: XL Repo: Fix availability tracker for plans.
When TopologyAvailabilityTracker is waiting for a plan topology, a "projected" topology for that plan will unblock waiters. I'm also changing the repository client to receive topology notifications from the start of the topic, so that we don't risk "losing" notifications on restart.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/34818/