OM-51594: Create Topology Extensions.
Problem
Create Topology Extensions and send Network Topology (Communication Matrix) using that mechanism.

Solution
1. Port Communication Matrix to XL
2. Modify Topology protobuf to include Topology Extensions
3. Make sure that nothing is broken at this stage.

The Topology Extensions:
The Topology Extension.
Due to the fact that Action Merging/Action Conversion is tied to a particular Topology, the need for Topology Extensions arises.
The Action Merging/Action Conversion is discovered by a probe, and is valid for a lifetime of a topology, which that particular discovery is a part of.
We have to preserve the Action Merging/Action Conversion specs, so that any Plans for that topology will generate same types of actions as a real-time Market.
By same type of action we mean that if actions would be merged/converted in a real-time Market, they would be merged/converted when running a Plan over the same Topology.
So, essentially, the Topology Extension is something that is tied to a Topology, has the lifetime of that Topology, yet not a part of Topology.
The Communication Matrix (Communication Topology) and Action Merging/Conversion data serve as examples.

The general idea is that some data associated with the topology and discovered by probes need to be available together with the topology.
The use cases would be real-time Market as well the Plans. The Plans based on a particular topology should have all of the same associated data, so that the actions produced during the plan would be coherent with the real-time Market.

In case of a Matrix (Network Topology), it is required by the Market for the placement decisions (whether move is actually a good idea).
In case of Action Merging/Conversion, it is needed so that that the kind of action produced (change config file instead of 10 scale actions) would be coherent with the real-time Market.

The receiving part is in a separate review.
Please refer to https://vmturbo.atlassian.net/wiki/spaces/Home/pages/1004962119/Action+Grouping+Architecture for details.

Testing
The unit tests have been augmented to send/receive Communication Matrix.
Full build with tests passed.
Added VC target and saw Market generates actions on my local.

RB: 34726