OM-51256, RB-34919: Enable SPEC2017 combined with 2006 in XL
Intent

Enable SPEC2017 combined with 2006 in XL so XL customers can use the latest CPUs only available in SPEC2017 as well as CPUs only in 2006.

Problem

Cisco wants to perform hardware refreshes using Gen 2 Intel cpu models. Unfortunately, none of those CPUs are in the spec 2006 benchmark. 73% of customers' cpu models are in the 2006 benchmark, but not 2017. As a result, we need to move to a system that uses both.

Solution

Use SPEC 2017 data combined with 2006 data.

Please see this review for a review of the solution
https://git.turbonomic.com/turbonomic/cpu-capacity/merge_requests/8

Related Improvement

Remove dependency on com.turbonomic.cpucapacity from com.vmturbo.api.component because api component talks to plan component using the cpu capacity service to get the data. cpucapacity project is only need on plan component which powers the cpu capacity service.