OM-47879 [RB-34833] Persist plan source topology in ArangoDB
Background
=======
We have been investigating ways to improve data consistency in plan data comparisons. The biggest gap currently is the source (or before) plan data, which is the source of confusion and apparent discrepancies. Our goal is to reduce and eliminate discrepancies by storing a snapshot of the plan source data with enough information to be able to display comparisons in a meaningful and consistent way.

Problem
=======
Most of the widgets on the Plan Results page request data from the realtime topology because there is not enough plan source topology data saved to serve their requests.

Solution
========
The Repository is modified so it both consumes and stores plan source topologies, allowing it to serve corresponding "before" data for all of the "after" data that it currently contains. Once the repository stores plan source data, we can change the widgets on the Plan Results page to request that data instead.