OM-51646, RB-35004: Read simplified component configuration on startup **background** Each Turbonomic component is designed to allow external configuration of nearly every customizable value via @Value injection during Spring configuration - typically injected via the Java System Properties settings. Further, the startup sequence, prior to Spring configuration, accesses the Sytem Properties settings to customize its behavior where those settings are required early on, e.g. the socket port number that the Jetty server will be initialized with. The prior solution for component configuration utilized ClusterMgr as a provider for the configuration properties for each component type. On startup, prior to Spring initialization, the component would pass the "component-default" properties to ClusterMgr to register the most recent defaults, and then in another request to ClusterMgr request the "effective" config properties - a combination of "default", "local", and "instance" properties.
**problem**
This prior solution requires two round-trip calls to ClusterMgr before the component can even begin configuration; this also requires ClustrMgr be up, which requires Auth, which required MariaDB.  This chain of dependencies adds significant complexity and delay to the startup.

**solution**
In OM.50042 we created a kubernetes config-map to be applied to each running component. This config map is mounted as a file, /etc/turbonomic/properties.yaml. This properties.yaml defines a map with two top-level sections: "defaultProperties" and "customProperties". Within defaultProperties are two types of subsections, "global" and "{component-type}", and each of those specifies a map with key/value pairs. Further, the "customProperties" may, optionally, define a "global" section or a "{component-type}" section with key/value pairs.
In other words, there are potentionally 4 sections of "properties.yaml" that determine the value of a property Y for a given component X:
1) defaultProperties/global/Y
2) defaultProperties/X/Y
3) customProperties/global/Y
4) customProperties/X/Y
These values are to be prioritized in this order, with the "last value defined wins".

Also in OM.50042 we enable setting config properties in the Custom Resource. This will allow easy configuration changes on a global or per-component basis, e.g.
```
properties:
   global:
     maxHealthyUsedMemoryRatio: 0.75
   topology_processor:
     broadcast_interval: 1min
```

In this commit we replace the ClusterMgr Configuration round-trip with code to simply read the "properties.yaml" mounted as /etc/turbonomic/properties.yaml and merge the different sections of key/value configuration as describe above. Then we apply the resulting key/value pairs to the System properties to be available for startup & Spring configuration

We also remove the "component_default.properties" file from each of the existing components, as this file will no longer be needed.