OM-52192-Fixed the issue that prevented system group for cloud vms with failed action gets created
Intent
======
Fix a number of issues that prevented system group for cloud vms with failed actions being created.

Problem
=======
We have the logic that auto creates a group named "Cloud VMs with Failed Sizing" if not present added vms that has a failed action to it. This group members gets updated as more vms actions get failed. However, the creation of that group fails.

Diagnosis
=========
There was a few issues that prevented this group to be created.
- The request for getting group did not have the property name which we are filtering based on.
- The create or update group grpc service did not allow creating a group where type has not been been set. THis is wrong since that property has a default value that should be used when the type is not set.
- The update method did not maintain the information about the origin of the group in the table.

Solution
========
I addressed the above issues as follows:
- I updated the request to have the property name included
- I changed the logic in group component to allow group definitions where type is not set.
- I changed the update the method to use existing values for user and system info when updating the group.