OM-51960, RB-35211: XL: Export diagnostics failed "Error message: Connect to rsyslog:8080"
background
There is a python script (diags.py) as part of the rsyslog container. This script opens an internal HTTP URL to export diagnostics files.

problem
This python script is written to Python 2.7 and the default for redhat ubi changed to python 3.

solution
Add "python2" to the list of packages to install in the Dockerfile for rsyslog