OM-51712: Completely reworked the QuoteCache class including interface.
# Motivation and Background

The old QuoteCache class was intentionally kept simple to leverage
library code as much as possible and prevent some classes of errors.
It used an implicit invalidation mechanism that -- although obviated the
need to review the code, find places where cache entries had to be
invalidated and then keep this set of places up-to-date -- resulted in
noticeable overhead in querying the cache. That was because the relevant
part of the trader state (its list of customers) had to be copied and
included in the key used to query the cache.

Since the underlying complexity of the shop-together algorithm isn't
affected by the cache (we still need to query the cache a quadratic
number of times) and since not all customer topologies have enough
biclique overlaps for caching to be profitable, we want to keep the
cache overhead to the absolute minimum.

Consequently, the hash map of triplets used in the old version of the
cache was replaced by a conceptual two-dimensional array and an explicit
invalidation mechanism. Indexing into this two-dimensional array is done
using integer indices that are already known at the time of the query,
minimising overhead. The design is actually a little more complex as an
additional array is used to translate between trader indices and the
actual row index into the two-dimensional array to prevent the latter
from getting too big.

Both the old and the new design waste memory in different ways (the
former by keeping invalidated cache entries in memory and the latter by
using a sparse array), but the amount of memory wasted by the new
approach is less than 2.5MB for a scenario bigger than our biggest
customers today.

# Changes

- Completely rewrote the QuoteCache class.
- Added a test suite for the QuoteCache class.
- Added documentation for all new production and test code.
- Updated `getMarketsAsBuyer` documentation to make the deterministic
  iteration order that was already implemented (and needed by the cache)
  part of the interface.
- Updated shop-together code to use the new cache and invalidate entries
  as needed.

Also:

- Fixed some nullability annotations that weren't in sync with code and
  were causing IDE warnings.
- Let IntelliJ IDEA rearrange imports to comply with standard.
- Made some fields `final`.

# Testing done

  - Run full test suite: All tests pass! That includes 205 brand-new
    unit tests. New code has 100% line and branch coverage.
  - At the time of this writing, the are no IntelliJ IDEA warnings for
    the new production code.
  - There are two IntelliJ IDEA warnings in new _test_ code about
    potential `NullPointerException`s.
      - The first appears to be due to the fact that IntelliJ IDEA
        doesn't know that the `get` method is deterministic, but I don't
        want to refactor the code to avoid this warning because if `get`
        ends up not being deterministic (currently this isn't checked
        statically), I want the test to fail.
      - The second points to a case where I actually want the test to
        fail with a `NullPointerException` if this ends up being `null`.
  - Reviewed most of the call tree rooted at
    `Trader#getModifiableCustomers` to make sure cache invalidation code
    is added at the appropriate places. Only some calls to `Move.take`
    and `Move.rollback` seemed appropriate.