OM-51000, RB-35191: Reduce memory footprint of calls to writeZipEntry
background
Each component implements a diagnostic dump method. This method is called by ClusterMgr via a REST api call, and the results are returned as a streaming HTTP response in zip-file format. The inner files within the zip are up to the component. The result stream for each component is then zipped into a zip output stream returned to API component, which then returns them to the end user.

problem
(1) The existing implementation creates a list of JSON strings and then passes the list to the diagnostic dump routine to be written to the output stream. This causes a huge memory spike for components that grow with large topologies - topology-processor, repository, action-orchestrator. This may also lead to health checks sent by kubernetes to fail (due to time out), and if enough failures occur the component is marked "down" and the long-running HTTP response socket is closed. The long-running HTTP response socket may also be closed if there is a sufficiently long idle time while the diagnostics gathers bytes to return.
(2) Individual objects, when converted to JSON, may be larger than 3MB:
repository-6d64fddff8-k5mvm: 2019-11-07 02:34:43,716 INFO [qtp12139181-13] [DiagnosticsWriter] : item 16405, total size: 3195277
Sending this JSON string as a single write leads to large buffer allocation further increasing memory demands.

solution
(1) add a new method to the diagnostic dump routine that takes a Stream<T> and uses the GSON mthod that serializes one T at a time directly to a stream, in this case the zip output stream.
(2) write large Strings to the zip output stream as chunks

Note that (2) would be unnecessary if all calls passed streams to the new method in (1). There are, however, cases where a wrapper object with an internal array is passed to be dumped. This will require a larger effort that should be included in this change. A new JIRA item will be added to capture this additional improvement.

Also, the methods to deserialize from a stream using GSON should be used instead of reading an entire file record into memory and then deserializing it. Another JIRA will be created for this improvement.