OM-51712: Modified QuoteCache to allow #sellers > #traders-in-economy.
Since there can't be more than nTradersInEconomy different trader
indices, it makes no sense to allocate space for more nPotentialSellers.
But, instead of throwing an `IllegalArgumentException`, the `QuoteCache`
will now internally use the lesser of the two sizes.

Since this wasn't as much of an error as a situation that could be
handled more efficiently, making this change should be safe. It is also
consistent with where and how `QuoteCache` is intended to be used.

Documentation and tests were updated as necessary. In total, the number
of tests was increased by 18.