OM-51712: Factored out a piece of cache invalidation code into a method.
A piece of code to invalidate the cache entries for both the source and
destination of a move was common between `CliqueMinimizer` and
`QuoteSummer` and was factored out into a different method.

Put the method into a different class as it's not considered part of the
cache interface itself and added documentation and tests. Only a minimal
set of tests were added (8) as the goal was to exercise the method
itself rather than repeating the tests in the `QuoteCache` test suite.
Line and branch coverage for the new code is 100%.