OM-52733: StatsAPI collapse multiple entities to one RB-35516 Issue ============
When scoping to the *Capacity and Usage* for a cloud Virtual Machine, commodity options are missing. Due this issues, VM capacity and usage values are not retrieved.

Background
============

@Roman devised a major portion of the code to fecth the correct stats using the relatedEntityType with the help of a new class called StatsQueryContext. In his words:

Instead of a "boolean" to represent the "full market" case, use a new GlobalScope helper object as part of the StatsQueryContext. The object will contain either a list of entities, or a set of restrictions on the global entities. Currently this set of restrictions is limited to related entity types, and environment types.

Problem
=========

The issue is happening in StatsQueryExecutor. The problem is that we make four sub-queries:

* CloudCostsStatsSubQuery
* RIStatsSubQuery
* StorageStatsSubQuery
* HistoricalCommodityStatsSubQuery
because no parameters are sent with the request. If I send a request with data:

```
{"statistics":[{"relatedEntityType":"VirtualMachine", "name":"test"}]}
```
then the commodity “test” is matched only for the HistoricalCommodity. That is a separate issue, but to solve this issue, we need a way to return one record using the current time, instead of all times.

Implementation
====================

If the input query does not specify a start or end time, we assume the user wants all commodities related to a specific time. Hence, I merge all the commodities into one stat and send it back.