OM-47880 [RB-35288] Allow API users to retrieve both plan source and plan projected data in a single call
Problem
=======
It is currently very difficult for the UI (or any API user) to compare plan source stats (the before) to plan projected stats (the after).

The current process goes something like this:

* Request the plan projected stats, get a page of results back
* Request the plan source stats for the specific entities in the first page of results
* Look at either actions or the supplychain to detect entities present in one topology and not the other. Make further requests to get their stats

The end result is the caller making 3+ API calls and then having to correlate the results. It is an error-prone process with many edge cases and performance implications.

Solution
========
The StatSnapshotApiDTO already supports returning multiple snapshots for each entity. If we combine the plan source and plan projected entities in the Repository (the data store), we can sort and paginate them together and return them in one response.

To enable this, we establish a convention to allow API users to indicate whether they want plan source data, plan projected data or both.

* If the requested start date is equal to (or earlier than) the plan run start time, we will include plan source data in the response.
* If the requested end date is greater than the plan run start time, we will include plan projected data in the response.