OM-52336,RB-35556-Determining the group cloud type based group entities rather than members in UUID mapper
Problem
=======
CCC widget in the scope of group of RGs is not being populated.

Diagnosis
=========
There is logic is uuidmapper which determines the type of cached groups. This logic is based on the members rather than entities. This does not work properly for nested groups as entities only have cloud type and groups don't. Since we cannot determine the cloud type of group we don't apply the cost query on that group which results in empty response.

Solution
========
Determine the cloud type of a group based on entities rather than members.