OM-52621, RB-35482: Allow Group Queries for BillingFamily in API
Intent
Allow group queries for BillingFamily in API so that Search->Billing Family UI can be changed to display the groups and their members instead of the master accounts.

Problem
API request for BillingFamily

Diagnosis
API needs to be aware of new Group types (Resource, BillingFamily, etc.)

Solution
Add Billing Family to GroupMapper's map and have SearchService consume that map. This allows us to only change two files (GroupFilterMapper and GroupMapper) when we add a new group type.

Alternative solutions
1. just add an additional if block and switch
    1. pro: simiplifies the review and code is straight forward
    2. con: unit test explosion having to cover all the group types
    3. con: every time we add a new group type, two location in api need to be changed
2. Do not populate businessAPI
    1. con: UI expects these values to be filled in. it won't work without
3. keep retrieval logic out of GroupMapper
    1. con: Would need to duplicate a lot of the code that populates the GroupApiDTO, or perform significant refactoring to reuse the GroupApiDTO code with the BillingFamilyApiDTO code

Other refactoring
I had to replace Immutable.of with a builder because guava does not support Immutable.of() with more than 10 parameters.