OM-51755 RB:35241 Backend changes to support developer freemium edition of Turbonomic.
Intention

To share the same backend code between XL and Lemur (The developer freemium edition of Turbonomic).
Lemur is targeted for application developers and will be provided for free in order to gain wider adoption of our software. Lemur is a trim downed version of XL, which only provides the visibility of entities and supply chains, with NO analysis and actions. Lemur is enabled with a special license with a "visibility_only" feature, and a limit of workload count of 100 (VMs).

Implementation

- Add check for developer freemium license feature.
- Implement developer freemuim REST interceptor to block REST endpoints that are not applicable.
- Refactor TopologyPipelineFactory into LivePipelineFactory and PlanPipeLineFactory, where LivePipelineFactory will construct different real time pipeline stages depending on the type of license.
- Disable market analysis when user applies a developer freemium license to a full XL installation.
- Wrap gRPC call in a try block and catch RuntimeException. If a gRPC call fails for whatever reason (including UnknownHostException), the gRPC framework will throw a io.grpc.StatusRuntimeException, and we will catch that to log a short meaningful warning message without stack trace. For other exceptions, log an error message with stack trace.
- Set default action severity to Normal when AO cannot be accessed.
- Remove ArangoHealthMonitor from Repository component
- Add auth.api dependency to the api and topology-processor profile